requirements.txt
streamlit==1.32.0
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
scikit-learn==1.4.1
requests==2.31.0